+++ game/missiongenerator/aircraft/aircraftbehavior.py
@@ -114,15 +114,15 @@ class AircraftBehavior:
         mission_uses_gun: bool = True,
         rtb_on_bingo: bool = True,
         ai_unlimited_fuel: Optional[bool] = None,
-        ai_no_gun: Optional[bool] = None,
+        ai_dont_use_gun: Optional[bool] = None,
     ) -> None:
         group.points[0].tasks.clear()
         if ai_unlimited_fuel is None:
             ai_unlimited_fuel = (
                 flight.squadron.coalition.game.settings.ai_unlimited_fuel
             )
-        if ai_no_gun is None:
-            ai_no_gun = flight.squadron.coalition.game.settings.ai_no_gun
+        if ai_dont_use_gun is None:
+            ai_dont_use_gun = flight.squadron.coalition.game.settings.ai_dont_use_gun
         # at IP, insert waypoint to orient aircraft in correct direction
         layout = flight.flight_plan.layout
         at_ip_or_combat = flight.state.is_at_ip or flight.state.in_combat
@@ -150,7 +150,7 @@ class AircraftBehavior:
         if restrict_jettison is not None:
             group.points[0].tasks.append(OptRestrictJettison(restrict_jettison))
         if rtb_winchester is not None:
-            if ai_no_gun and flight.is_helo and flight.client_count == 0:
+            if ai_dont_use_gun and flight.is_helo:
                 group.points[0].tasks.append(
                     OptRTBOnOutOfAmmo(OptRTBOnOutOfAmmo.Values.Rockets)
                 )
@@ -212,13 +212,6 @@ class AircraftBehavior:
             ammo_type = OptRTBOnOutOfAmmo.Values.Cannon
 
         self.configure_behavior(flight, group, rtb_winchester=ammo_type)
-
-    def configure_ai_no_gun(
-        self, group: FlyingGroup[Any], flight: Flight, ai_no_gun: bool = False
-    ) -> None:
-        if ai_no_gun and flight.is_helo and flight.client_count == 0:
-            for unit in group.units:
-                unit.gun = 0
 
     def configure_cas(self, group: FlyingGroup[Any], flight: Flight) -> None:
         self.configure_task(flight, group, CAS, [AFAC, AntishipStrike])